Feature: 파일명 Rename, uploadSchedulesToServer 구현
